requirements.txt
# Core dependencies
requests>=2.31.0
neo4j>=5.14.0
openai>=1.0.0

# NLP and ML
gliner2>=0.1.0

# UI
gradio>=4.0.0
pyvis>=0.3.2
networkx>=3.0

# Development
pytest>=7.4.0
mypy>=1.7.0
python-dotenv>=1.0.0
types-requests>=2.31.0
